requirements.txt
cython==0.19.2
numpy==1.8.0
biopython==1.62
matplotlib==1.3.1
intermine==1.09.05
